Rename/repurpose the previous experiment
Starting to toy with the idea of making this into something that will work.
